+++ lib/spectre.rb
@@ -1,11 +1,31 @@
 module Spectre
   module Version
     MAJOR = 1
-    MINOR = 11
+    MINOR = 12
     TINY  = 0
   end
 
   VERSION = [Version::MAJOR, Version::MINOR, Version::TINY].compact * '.'
+
+
+  class ::Hash
+    def deep_merge!(second)
+      merger = proc { |key, v1, v2| Hash === v1 && Hash === v2 ? v1.merge!(v2, &merger) : v2 }
+      self.merge!(second, &merger)
+    end
+
+    def deep_clone
+      Marshal.load(Marshal.dump(self))
+    end
+  end
+
+  class ::Object
+    def to_h
+      self.instance_variables.each_with_object({}) do |var, hash|
+        hash[var.to_s.delete("@")] = self.instance_variable_get(var)
+      end
+    end
+  end
 
 
   ###########################################
@@ -35,9 +55,9 @@ module Spectre
       instance_eval(&block)
     end
 
-    def _execute args, &block
+    def _execute *args, &block
       @__bound_self__ = eval('self', block.binding)
-      instance_exec(args, &block)
+      instance_exec(*args, &block)
     end
 
     def method_missing method, *args, **kwargs, &block
@@ -48,7 +68,6 @@ module Spectre
       end
     end
   end
-
 
   class Subject
     attr_reader :name, :desc, :specs
@@ -64,7 +83,6 @@ module Spectre
       @specs << Spec.new(name, self, desc, tags, data, block, context, file)
     end
   end
-
 
   class Spec
     attr_reader :name, :subject, :context, :desc, :tags, :data, :block, :file
@@ -84,7 +102,6 @@ module Spectre
       @subject.desc + ' ' + desc
     end
   end
-
 
   class RunInfo
     attr_accessor :spec, :data, :started, :finished, :error, :failure, :skipped, :log, :properties
@@ -113,7 +130,6 @@ module Spectre
       @error != nil
     end
   end
-
 
   class Runner
     @@current
@@ -151,11 +167,15 @@ module Spectre
 
       begin
         specs.each do |spec|
+          raise "Multi data definition (`with' parameter) of '#{spec.subject.desc} #{spec.desc}' has to be an `Array'" unless !spec.data.nil? and spec.data.is_a? Array
+
           if spec.data.any?
-            spec.data.each do |data|
-              Logger.log_spec(spec, data) do
-                runs << run_spec(spec, data)
-              end
+            spec.data
+              .map { |x| x.is_a?(Hash) ? OpenStruct.new(x) : x }
+              .each do |data|
+                Logger.log_spec(spec, data) do
+                  runs << run_spec(spec, data)
+                end
             end
           else
             Logger.log_spec(spec) do
@@ -189,14 +209,11 @@ module Spectre
           end
 
           run_info.finished = Time.now
-
         rescue ExpectationFailure => e
           run_info.failure = e
-
         rescue Exception => e
           run_info.error = e
           Logger.log_error spec, e
-
         end
       end
 
@@ -226,18 +243,14 @@ module Spectre
         end
 
         spec.block.call(data)
-
       rescue ExpectationFailure => e
         run_info.failure = e
-
       rescue Interrupt
         run_info.skipped = true
         Logger.log_skipped spec
-
       rescue Exception => e
         run_info.error = e
         Logger.log_error spec, e
-
       ensure
         if spec.context.__after_blocks.count > 0
           after_ctx = SpecContext.new(spec.subject, 'after')
@@ -249,14 +262,11 @@ module Spectre
               end
 
               run_info.finished = Time.now
-
             rescue ExpectationFailure => e
               run_info.failure = e
-
             rescue Exception => e
               run_info.error = e
               Logger.log_error spec, e
-
             end
           end
         end
@@ -290,7 +300,6 @@ module Spectre
     end
 
     def it desc, tags: [], with: [], &block
-
       # Get the file, where the spec is defined.
       # Nasty, but it works
       # Maybe there is another way, but this works for now
@@ -306,11 +315,7 @@ module Spectre
           .first
       end
 
-      raise "`with' has to be an Array" unless with.is_a? Array
-
-      data = with.map { |x| x.is_a?(Hash) ? OpenStruct.new(x) : x }
-
-      @__subject.add_spec(desc, tags, data, block, self, spec_file)
+      @__subject.add_spec(desc, tags, with, block, self, spec_file)
     end
 
     def before &block
@@ -348,6 +353,7 @@ module Spectre
       methods.each do |method_name|
         define_method(method_name) do |*args, &block|
           return super(*args, &block) if respond_to? method_name
+
           target.send(method_name, *args, &block)
         end
 
@@ -358,8 +364,9 @@ module Spectre
     end
 
     def self.redirect method_name, *args, **kwargs, &block
-      target = @@mappings[method_name]
-      raise "No method or variable '#{method_name}' defined" if !target
+      target = @@mappings[method_name] || Kernel
+      raise "no variable or method '#{method_name}' found" unless target.respond_to? method_name
+
       target.send(method_name, *args, **kwargs, &block)
     end
   end
@@ -377,12 +384,11 @@ module Spectre
         .map { |x| x.specs }
         .flatten
         .select do |spec|
-          (spec_filter.empty? or spec_filter.any? { |x| spec.name.match('^' + x.gsub('*', '.*') + '$') }) and (tags.empty? or tags.any? { |x| has_tag(spec.tags, x) })
+          (spec_filter.empty? or spec_filter.any? { |x| spec.name.match('^' + x.gsub('*', '.*') + '$') }) and (tags.empty? or tags.any? { |x| tag?(spec.tags, x) })
         end
     end
 
-
-    def has_tag tags, tag_exp
+    def tag? tags, tag_exp
       tags = tags.map { |x| x.to_s }
       all_tags = tag_exp.split '+'
       included_tags = all_tags.select { |x| !x.start_with? '!' }
@@ -390,16 +396,13 @@ module Spectre
       included_tags & tags == included_tags and excluded_tags & tags == []
     end
 
-
     def delegate *method_names, to: nil
       Spectre::Delegator.delegate *method_names, to
     end
 
-
     def register &block
       @@modules << block
     end
-
 
     def configure config
       @@modules.each do |block|
@@ -416,7 +419,7 @@ module Spectre
     def describe desc, &block
       subject = @@subjects.find { |x| x.desc == desc }
 
-      if !subject
+      unless subject
         subject = Subject.new(desc)
         @@subjects << subject
       end
@@ -428,7 +431,6 @@ module Spectre
     def property key, val
       Spectre::Runner.current.properties[key] = val
     end
-
   end
 
   delegate :describe, :property, to: Spectre