+++ lib/spectre/curl.rb
@@ -1,3 +1,5 @@
+require_relative '../spectre'
+
 require 'open3'
 require 'ostruct'
 
@@ -21,17 +23,17 @@ module Spectre::Curl
     end
 
     def header name, value
-      @__req['headers'] = [] if not @__req['headers']
+      @__req['headers'] = [] unless @__req['headers']
       @__req['headers'].append [name, value.to_s.strip]
     end
 
     def param name, value
-      @__req['query'] = [] if not @__req['query']
+      @__req['query'] = [] unless @__req['query']
       @__req['query'].append [name, value.to_s.strip]
     end
 
     def content_type media_type
-      @__req['headers'] = [] if not @__req['headers']
+      @__req['headers'] = [] unless @__req['headers']
       @__req['headers'].append ['Content-Type', media_type]
     end
 
@@ -111,7 +113,7 @@ module Spectre::Curl
     end
 
     def json
-      return nil if not @res[:body]
+      return nil unless @res[:body]
 
       if @data == nil
         begin
@@ -161,28 +163,31 @@ module Spectre::Curl
 
       if @@http_cfg.key? name
         req.merge! @@http_cfg[name]
-        raise "No `base_url' set for HTTP client '#{name}'. Check your HTTP config in your environment." if !req['base_url']
+        raise "No `base_url' set for HTTP client '#{name}'. Check your HTTP config in your environment." unless req['base_url']
       else
         req['base_url'] = name
       end
 
-      SpectreHttpRequest.new(req).instance_eval(&block) if block_given?
+      SpectreHttpRequest.new(req)._evaluate(&block) if block_given?
 
       invoke(req)
     end
 
     def curl_request
       raise 'No request has been invoked yet' unless @@request
+
       @@request
     end
 
     def curl_response
       raise 'There is no response. No request has been invoked yet.' unless @@response
+
       @@response
     end
 
     def register mod
       raise 'Module must not be nil' unless mod
+
       @@modules << mod
     end
 
@@ -193,7 +198,7 @@ module Spectre::Curl
 
       begin
         json = JSON.parse(str)
-        json.obfuscate!(@@secure_keys) if not @@debug
+        json.obfuscate!(@@secure_keys) unless @@debug
 
         if pretty
           str = JSON.pretty_generate(json)
@@ -207,8 +212,8 @@ module Spectre::Curl
       str
     end
 
-    def is_secure? key
-      @@secure_keys.any? { |x| key.to_s.downcase.include? x.downcase  }
+    def secure? key
+      @@secure_keys.any? { |x| key.to_s.downcase.include? x.downcase }
     end
 
     def header_to_s headers
@@ -219,7 +224,7 @@ module Spectre::Curl
       headers.each do |header|
         key = header[0].to_s
         value = header[1].to_s
-        value = '*****' if is_secure?(key) and not @@debug
+        value = '*****' if secure?(key) and not @@debug
         s += "#{key.ljust(30, '.')}: #{value}\n"
       end
 
@@ -237,12 +242,12 @@ module Spectre::Curl
 
       uri = req['base_url']
 
-      if not uri.match /http(?:s)?:\/\//
+      unless uri.match /http(?:s)?:\/\//
         uri = scheme + '://' + uri
       end
 
       if req['path']
-        uri += '/' if !uri.end_with? '/'
+        uri += '/' unless uri.end_with? '/'
         uri += req['path']
       end
 
@@ -277,6 +282,7 @@ module Spectre::Curl
       # Add certificate path if one if given
       if req['cert']
         raise "Certificate '#{req['cert']}' does not exist" unless File.exists? req['cert']
+
         cmd.append('--cacert', req['cert'])
       elsif req['use_ssl'] or uri.start_with? 'https'
         cmd.append('-k')
@@ -313,7 +319,7 @@ module Spectre::Curl
 
       # debug_log.lines.each { |x| @@logger.debug x unless x.empty? }
 
-      raise "Unable to request #{uri}. Please check if this service is reachable." unless output
+      raise "Unable to request #{uri}. Please check if this URL is correctly configured and reachable." unless output
 
       @@logger.debug("[<] #{req_id} stdout:\n#{output}")
 
@@ -341,7 +347,7 @@ module Spectre::Curl
         code: match[:code].to_i,
         message: match[:message],
         headers: Hash[res_headers],
-        body: body
+        body: body,
       }
 
       # Call all registered modules