+++ lib/spectre/environment.rb
@@ -1,9 +1,13 @@
+require_relative '../spectre'
+
 require 'ostruct'
 
 def to_recursive_ostruct(hash)
-  OpenStruct.new(hash.each_with_object({}) do |(key, val), memo|
-    memo[key] = val.is_a?(Hash) ? to_recursive_ostruct(val) : val
-  end)
+  OpenStruct.new(
+    hash.each_with_object({}) do |(key, val), memo|
+      memo[key] = val.is_a?(Hash) ? to_recursive_ostruct(val) : val
+    end
+  )
 end
 
 module Spectre
@@ -21,6 +25,6 @@ module Spectre
       @@environment.freeze
     end
 
-    Spectre.delegate :env, to: Environment
+    Spectre.delegate :env, to: self
   end
 end