+++ lib/spectre/http.rb
@@ -1,10 +1,11 @@
+require_relative '../spectre'
+
 require 'net/http'
 require 'openssl'
 require 'json'
 require 'securerandom'
 require 'logger'
 require 'ostruct'
-
 
 module Spectre
   module Http
@@ -19,11 +20,16 @@ module Spectre
       'headers' => nil,
       'query' => nil,
       'content_type' => '',
-    }
+    }.freeze
 
     @@modules = []
 
     class SpectreHttpRequest < Spectre::DslClass
+      class Headers
+        CONTENT_TYPE = 'Content-Type'
+        UNIQUE_HEADERS = [CONTENT_TYPE].freeze
+      end
+
       def initialize request
         @__req = request
       end
@@ -40,25 +46,30 @@ module Spectre
         @__req['path'] = url_path
       end
 
+      def timeout seconds
+        @__req['timeout'] = seconds
+      end
+
       def header name, value
-        @__req['headers'] = [] if not @__req['headers']
+        @__req['headers'] ||= []
         @__req['headers'].append [name, value.to_s.strip]
       end
 
       def param name, value
-        @__req['query'] = [] if not @__req['query']
+        @__req['query'] ||= []
         @__req['query'].append [name, value.to_s.strip]
       end
 
       def content_type media_type
-        @__req['headers'] = [] if not @__req['headers']
-        @__req['headers'].append ['Content-Type', media_type]
+        @__req['content_type'] = media_type
       end
 
       def json data
         data = data.to_h if data.is_a? OpenStruct
         body JSON.pretty_generate(data)
-        content_type 'application/json'
+
+        # TODO: Only set content type, if not explicitly set
+        content_type('application/json')
       end
 
       def body body_content
@@ -77,9 +88,12 @@ module Spectre
         @__req['auth'] = method
       end
 
+      def no_auth!
+        @__req['auth'] = 'none'
+      end
+
       def certificate path
         @__req['cert'] = path
-        use_ssl!
       end
 
       def use_ssl!
@@ -101,7 +115,8 @@ module Spectre
       end
 
       def [] key
-        return nil if not @headers.has_key?(key.downcase)
+        return nil unless @headers.key?(key.downcase)
+
         @headers[key.downcase].first
       end
 
@@ -111,52 +126,30 @@ module Spectre
     end
 
     class SpectreHttpResponse
-      def initialize res
-        @res = res
-        @data = nil
-        @headers = SpectreHttpHeader.new @res[:headers]
-      end
+      attr_reader :code, :message, :headers, :body
 
-      def code
-        @res[:code]
-      end
-
-      def message
-        @res[:message]
-      end
-
-      def headers
-        @headers
-      end
-
-      def body
-        @res[:body]
+      def initialize net_res
+        @code = net_res.code.to_i
+        @message = net_res.message
+        @body = net_res.body
+        @headers = SpectreHttpHeader.new(net_res.to_hash)
+        @json_data = nil
       end
 
       def json
-        return nil if not @res[:body]
-
-        if @data == nil
+        if !@body.nil? and @json_data.nil?
           begin
-            @data = JSON.parse(@res[:body], object_class: OpenStruct)
-          rescue
-            raise "Body content is not a valid JSON:\n#{@res[:body]}"
+            @json_data = JSON.parse(@body, object_class: OpenStruct)
+          rescue JSON::ParserError
+            raise "Body content is not a valid JSON:\n#{@body}"
           end
         end
 
-        @data
+        @json_data
       end
 
       def success?
-        @res[:code] < 400
-      end
-
-      def to_s
-        @res.to_s
-      end
-
-      def pretty
-        @res.pretty
+        @code < 400
       end
     end
 
@@ -172,35 +165,38 @@ module Spectre
         http(name, secure: true, &block)
       end
 
-      def http name, secure: nil, &block
+      def http name, secure: false, &block
         req = {}
 
         if @@http_cfg.key? name
-          req.merge! @@http_cfg[name]
-          raise "No `base_url' set for HTTP client '#{name}'. Check your HTTP config in your environment." if !req['base_url']
+          req.deep_merge! @@http_cfg[name].deep_clone
+          raise "No `base_url' set for HTTP client '#{name}'. Check your HTTP config in your environment." unless req['base_url']
         else
           req['base_url'] = name
         end
 
-        req['user_ssl'] = secure if secure != nil
+        req['use_ssl'] = secure unless secure.nil?
 
-        SpectreHttpRequest.new(req).instance_eval(&block) if block_given?
+        SpectreHttpRequest.new(req)._evaluate(&block) if block_given?
 
         invoke(req)
       end
 
       def request
         raise 'No request has been invoked yet' unless @@request
+
         @@request
       end
 
       def response
         raise 'There is no response. No request has been invoked yet.' unless @@response
+
         @@response
       end
 
       def register mod
         raise 'Module must not be nil' unless mod
+
         @@modules << mod
       end
 
@@ -211,7 +207,7 @@ module Spectre
 
         begin
           json = JSON.parse(str)
-          json.obfuscate!(@@secure_keys) if not @@debug
+          json.obfuscate!(@@secure_keys) unless @@debug
 
           if pretty
             str = JSON.pretty_generate(json)
@@ -225,14 +221,14 @@ module Spectre
         str
       end
 
-      def is_secure? key
-        @@secure_keys.any? { |x| key.to_s.downcase.include? x.downcase  }
+      def secure? key
+        @@secure_keys.any? { |x| key.to_s.downcase.include? x.downcase }
       end
 
       def header_to_s headers
         s = ''
         headers.each_header.each do |header, value|
-          value = '*****' if is_secure?(header) and not @@debug
+          value = '*****' if secure?(header) and not @@debug
           s += "#{header.to_s.ljust(30, '.')}: #{value.to_s}\n"
         end
         s
@@ -241,42 +237,47 @@ module Spectre
       def invoke req
         @@request = nil
 
-        if req['cert'] or req['use_ssl']
-          scheme = 'https'
-        else
-          scheme = 'http'
-        end
+        # Build URI
 
+        scheme = req['use_ssl'] ? 'https' : 'http'
         base_url = req['base_url']
 
-        if not base_url.match /http(?:s)?:\/\//
+        unless base_url.match /http(?:s)?:\/\//
           base_url = scheme + '://' + base_url
         end
 
         if req['path']
-          base_url = base_url + '/' if not base_url.end_with? '/'
+          base_url = base_url + '/' unless base_url.end_with? '/'
           base_url += req['path']
         end
 
         uri = URI(base_url)
 
-        raise "'#{uri}' is not a valid uri" if not uri.host
+        raise "'#{uri}' is not a valid uri" unless uri.host
+
+        # Build query parameters
 
         uri.query = URI.encode_www_form(req['query']) unless not req['query'] or req['query'].empty?
 
+        # Create HTTP client
+
         net_http = Net::HTTP.new(uri.host, uri.port)
+        net_http.read_timeout = req['timeout'] || 180
 
         if uri.scheme == 'https'
           net_http.use_ssl = true
 
           if req.key? 'cert'
             raise "Certificate '#{req['cert']}' does not exist" unless File.exists? req['cert']
+
             net_http.verify_mode = OpenSSL::SSL::VERIFY_PEER
             net_http.ca_file = req['cert']
           else
             net_http.verify_mode = OpenSSL::SSL::VERIFY_NONE
           end
         end
+
+        # Create HTTP Request
 
         net_req = Net::HTTPGenericRequest.new(req['method'], true, true, uri)
         net_req.body = req['body']
@@ -302,13 +303,22 @@ module Spectre
         req_log += header_to_s(net_req)
         req_log += try_format_json(req['body'], pretty: true) if req['body'] != nil and not req['body'].empty?
 
-        @@logger.info req_log
+        @@logger.info(req_log)
 
         # Request
 
         start_time = Time.now
-        net_res = net_http.request(net_req)
+
+        begin
+          net_res = net_http.request(net_req)
+        rescue SocketError => e
+          raise "The request '#{req['method']} #{uri}' failed. Please check if the given URL '#{uri}' is valid and available or a corresponding HTTP config in the environment file exists. See log for more details. Original.\nOriginal error was: #{e.message}"
+        end
+
         end_time = Time.now
+
+        req['started_at'] = start_time
+        req['finished_at'] = end_time
 
         # Run HTTP modules
 
@@ -320,27 +330,23 @@ module Spectre
 
         res_log = "[<] #{req_id} #{net_res.code} #{net_res.message} (#{end_time - start_time}s)\n"
         res_log += header_to_s(net_res)
-        res_log += try_format_json(net_res.body, pretty: true) if net_res.body != nil and !net_res.body.empty?
+        res_log += try_format_json(net_res.body, pretty: true) unless net_res.body.nil? or net_res.body.empty?
 
         @@logger.info(res_log)
 
-        if req['ensure_success']
-          code = Integer(net_res.code)
-          fail "Response code of #{req_id} did not indicate success: #{net_res.code} #{net_res.message}" if code >= 400
-        end
+        fail "Response code of #{req_id} did not indicate success: #{net_res.code} #{net_res.message}" if req['ensure_success'] and net_res.code.to_i >= 400
+
+        # Set global request and response variables
 
         @@request = OpenStruct.new(req)
-        @@response = SpectreHttpResponse.new({
-          code: net_res.code.to_i,
-          message: net_res.message,
-          headers: net_res.to_hash,
-          body: net_res.body
-        })
+        @@request.freeze
+
+        @@response = SpectreHttpResponse.new(net_res)
       end
     end
 
     Spectre.register do |config|
-      @@logger = ::Logger.new config['log_file'], progname: 'spectre/http'
+      @@logger = ::Logger.new(config['log_file'], progname: 'spectre/http')
       @@secure_keys = config['secure_keys'] || []
       @@debug = config['debug']