+++ lib/spectre/mixin.rb
@@ -1,8 +1,27 @@
+require_relative '../spectre'
+require_relative 'logger'
+
 require 'ostruct'
-require 'spectre/logger'
 
 module Spectre
   module Mixin
+    class MixinContext < Spectre::DslClass
+      def initialize desc
+        @__desc = desc
+      end
+
+      def required params, *keys
+        missing_keys = keys.select { |x| !params.to_h.key? x }
+        Spectre::Logger.log_debug("required parameters for '#{@__desc}': #{keys.join ', '}")
+        raise ArgumentError, "mixin '#{@__desc}' requires #{keys.join ', '}, but only has #{missing_keys.join ', '} given" unless missing_keys.empty?
+      end
+
+      def optional params, *keys
+        Spectre::Logger.log_debug("optional parameters for '#{@__desc}': #{keys.join ', '}")
+        params
+      end
+    end
+
     class << self
       @@mixins = {}
 
@@ -12,23 +31,28 @@ module Spectre
 
       def run desc, with: []
         raise "no mixin with desc '#{desc}' defined" unless @@mixins.key? desc
+
         Spectre::Logger.log_debug "running mixin '#{desc}'"
 
         params = with || {}
 
+        ctx = MixinContext.new(desc)
+
         if params.is_a? Array
-          @@mixins[desc].call *params
+          return_val = ctx._execute(*params, &@@mixins[desc])
         elsif params.is_a? Hash
-          @@mixins[desc].call OpenStruct.new(params)
+          return_val = ctx._execute(OpenStruct.new(params), &@@mixins[desc])
         else
-          @@mixins[desc].call params
+          return_val = ctx._execute(params, &@@mixins[desc])
         end
+
+        return_val.is_a?(Hash) ? OpenStruct.new(return_val) : return_val
       end
 
       alias_method :also, :run
       alias_method :step, :run
     end
 
-    Spectre.delegate :mixin, :run, :also, :step, to: Mixin
+    Spectre.delegate :mixin, :run, :also, :step, to: self
   end
 end