+++ lib/spectre/resources.rb
@@ -1,3 +1,5 @@
+require_relative '../spectre'
+
 require 'ostruct'
 
 module Spectre
@@ -12,7 +14,8 @@ module Spectre
       end
 
       def [] name
-        raise "Resource with name '#{name}' does not exist" if not @items.key? name
+        raise "Resource with name '#{name}' does not exist" unless @items.key? name
+
         @items[name]
       end
     end
@@ -26,7 +29,7 @@ module Spectre
     end
 
     Spectre.register do |config|
-      return if !config.key? 'resource_paths'
+      return unless config.key? 'resource_paths'
 
       config['resource_paths'].each do |resource_path|
         resource_files = Dir.glob File.join(resource_path, '**/*')
@@ -41,6 +44,6 @@ module Spectre
       @@resources.freeze
     end
 
-    Spectre.delegate :resources, to: Resources
+    Spectre.delegate :resources, to: self
   end
 end